Change type value

+++ processcontrol_fields/models/sale_order_line.py
@@ -8,6 +8,6 @@ from odoo.exceptions import ValidationError
 class SaleOrderLine(models.Model):
     _inherit = ['sale.order.line']
 
-    horas_reales = fields.Integer(
+    horas_reales = fields.Float(
         string='Horas Reales',
         required=False)
